Move the simple binary editing functions from BinaryFontEditor to BinaryEditor.
